python-api/requirements.txt
numpy==1.26.4
python-dotenv==1.0.1
setuptools==69.2.0
robyn==0.53.1
pillow>=9.0.0.post1
onnxruntime==1.17.1